feature(zone): kernel command line control on launch (#351)

+++ crates/xen/xencall/examples/console_read.rs
@@ -6,7 +6,7 @@ async fn main() -> Result<()> {
     env_logger::init();
 
     let call = XenCall::open(0)?;
-    let index = 0 as u32;
+    let index = 0_u32;
     let (buf, newindex) = call.read_console_ring_raw(false, index).await?;
 
     match std::str::from_utf8(&buf[..newindex as usize]) {
